python/requirements.txt
aws-cdk-lib>=2.10.0
constructs>=10.0.60